requirements.txt
docutils
ipykernel
packaging
tornado>=6.1.0
jupyter-core
jupyter-packaging
jupyterlab-server>=2.11.1,<3
jupyter-server>=1.16.0,<2
notebook_shim>=0.1
jinja2>=3.0.3
jupyterlab==3.4.6
jupyterlab-widgets==1.1.0
nbformat
numpy
requests
gitpython
pygithub
pyunicore
siibra
tqdm
tvb-library
tvb-gdist==2.1.0
tvb-framework
tvb-ext-bucket
tvb-ext-unicore
tvb-widgets>=1.0